+++ src/modal/DetailModal.ts
